add updateMPC

+++ source/controllers/mpc/gait.cc
@@ -2,7 +2,7 @@
 
 namespace sdrobot::ctrl::mpc
 {
-  OffsetDurationGait::OffsetDurationGait(int nSegment, Eigen::Vector4i offset, Eigen::Vector4i durations, const std::string &name)
+  OffsetDurationGait::OffsetDurationGait(unsigned nSegment, Eigen::Vector4i offset, Eigen::Vector4i durations, const std::string &name)
   {
   }
   Vector4 OffsetDurationGait::getContactState() { return Vector4(); }
@@ -11,11 +11,11 @@ namespace sdrobot::ctrl::mpc
 
   const std::vector<int> &OffsetDurationGait::getMpcTable() { return _mpc_table; }
 
-  void OffsetDurationGait::setIterations([[maybe_unused]] int iterationsBetweenMPC, [[maybe_unused]] int currentIteration) { return; }
+  void OffsetDurationGait::setIterations([[maybe_unused]] unsigned iterationsBetweenMPC, [[maybe_unused]] unsigned currentIteration) { return; }
 
-  double OffsetDurationGait::getCurrentStanceTime(double dtMPC, int leg) { return dtMPC + leg; }
+  double OffsetDurationGait::getCurrentStanceTime(double dtMPC, size_t leg) { return dtMPC + leg; }
 
-  double OffsetDurationGait::getCurrentSwingTime(double dtMPC, int leg) { return dtMPC + leg; }
+  double OffsetDurationGait::getCurrentSwingTime(double dtMPC, size_t leg) { return dtMPC + leg; }
 
-  int OffsetDurationGait::getCurrentGaitPhase() { return 0; }
+  unsigned OffsetDurationGait::getCurrentGaitPhase() { return 0; }
 }
